requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
ansiwrap==0.8.4
appdirs==1.4.4
async-generator==1.10; python_version >= '3.5'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
bcrypt==3.2.0; python_version >= '3.6'
black==21.5b2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.1; python_version >= '3.6'
cryptography==3.4.7; python_version >= '3.6'
dependency-injector==4.32.3
entrypoints==0.3; python_version >= '2.7'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-client==6.1.12; python_version >= '3.5'
jupyter-core==4.7.1; python_version >= '3.6'
mypy-extensions==0.4.3
nbclient==0.5.3; python_full_version >= '3.6.1'
nbformat==5.1.3; python_version >= '3.5'
nest-asyncio==1.5.1; python_version >= '3.5'
numpy==1.20.3
pandas==1.2.4
papermill==2.3.3
paramiko==2.7.2
pathspec==0.8.1
pyarrow==4.0.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.8.2
pymysql==1.0.2
pynacl==1.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.17.3; python_version >= '3.5'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.17.1
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.1.0; python_version >= '3.6'
regex==2021.4.4
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sshtunnel==0.4.0
tenacity==7.0.0
textwrap3==0.9.2
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.1; python_version >= '3.5'
tqdm==4.61.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.0.5; python_version >= '3.7'
typing-extensions==3.10.0.0
urllib3==1.26.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
